chore: rename to bio
